Clean the landscape plotting utilities
