+++ .gitignore
@@ -36,6 +36,9 @@ htmlcov
 # Pictures created by backtesting
 *.png
 
+# Picture components for documentation
+docs/scripts/graphics/*.svg
+
 # Folders that are temporarily created when building the documentation
 docs/_autosummary
 docs/_build
